+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - ID指定されたイベントについてはEventScheduledでなくても予約集計するように調整
 - 予約検索にcountDocuments指定を追加
 - 予約にuseActionExists属性を追加
+- 予約にreservedTicket.dateUsed属性を実験的に追加
 - 予約検索に$projectionを指定できるように調整
 - 予約使用時にagent.identifierを指定できるように調整